% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/josa.R
\name{josa}
\alias{josa}
\title{josa support}
\usage{
josa(textko, select)
}
\arguments{
\item{textko}{target text in Korean encoding UTF-8}

\item{select}{Joso that need to select.}
}
\description{
josa support
}
\examples{
 josa("텀블러", "와")
}
