% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/josaplay-package.R
\docType{package}
\name{josaplay-package}
\alias{josaplay}
\alias{josaplay-package}
\title{josaplay: Add Josa Based on Previous Letter in Korean}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Josa in Korean is often determined by judging the previous word. When writing reports using Rmd, a function that prints the appropriate investigation for each case is helpful. The 'josaplay' package then evaluates the previous word to determine which josa is appropriate.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrchypark/josaplay}
  \item Report bugs at \url{https://github.com/mrchypark/josaplay/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (\href{https://orcid.org/0000-0001-6474-2570}{ORCID})

}
\keyword{internal}
