% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/josaplay-package.R
\docType{package}
\name{josaplay-package}
\alias{josaplay}
\alias{josaplay-package}
\title{josaplay: Add Josa Based on Previous Letter in Korean}
\description{
Add josa text for Korean.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrchypark/josaplay}
  \item Report bugs at \url{https://github.com/mrchypark/josaplay/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (0000-0001-6474-2570)

}
\keyword{internal}
