% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsp.R
\name{jsp_gwawa}
\alias{jsp_gwawa}
\alias{jsp_eunneun}
\alias{jsp_iga}
\alias{jsp_eulleul}
\title{josa support}
\arguments{
\item{textko}{target text in Korean encoding UTF-8}
}
\description{
prefix jsp_ functions work support gwawa, eunneun, iga, eulleul
}
\examples{
 jsp_gwawa("134")
 jsp_eulleul(15)
}
