% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsp.R
\name{extract}
\alias{extract}
\alias{jsp_gwawa}
\alias{jsp_eunneun}
\alias{jsp_iga}
\alias{jsp_eulleul}
\title{josa support (legacy)}
\usage{
jsp_gwawa(textko)

jsp_eunneun(textko)

jsp_iga(textko)

jsp_eulleul(textko)
}
\arguments{
\item{textko}{target text in Korean encoding UTF-8}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
prefix jsp_ functions work support gwawa, eunneun, iga, eulleul
}
\examples{
 jsp_gwawa("134")
 jsp_eulleul(15)
}
